+++ spec/lib/simp/rpm_spec.rb
@@ -3,7 +3,6 @@ require 'spec_helper'
 
 describe Simp::RPM do
   before :all do
-    Simp::RPM.stubs(:system_dist).returns('.testdist')
 
     dir          = File.expand_path( 'files', File.dirname( __FILE__ ) )
     @spec_file   = File.join( dir, 'testpackage.spec' )
@@ -225,10 +224,16 @@ describe Simp::RPM do
 
     context '#dist' do
       it 'returns dist' do
-        expect( @rpm_obj.dist ).to eq '.testdist'
-        expect( @d_rpm_obj.dist ).to eq '.el7'
-        expect( @spec_obj.dist ).to eq '.testdist'
-        expect( @d_spec_obj.dist ).to eq '.testdist'
+        Simp::RPM.stubs(:system_dist).returns('.testdist')
+        rpm_obj    = Simp::RPM.new( @rpm_file )
+        d_rpm_obj  = Simp::RPM.new( @d_rpm_file )
+        spec_obj   = Simp::RPM.new( @spec_file )
+        d_spec_obj = Simp::RPM.new( @d_spec_file )
+
+        expect( rpm_obj.dist ).to eq '.testdist'
+        expect( d_rpm_obj.dist ).to eq '.el7'
+        expect( spec_obj.dist ).to eq '.testdist'
+        expect( d_spec_obj.dist ).to eq '.testdist'
       end
 
       it 'fails when invalid package specified' do