+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 5.5.0 / 2018-03-30
+* Add support for setting SIMP_RSPEC_PUPPETFILE and/or SIMP_RSPEC_MODULEPATH to
+  create a custom fixtures.yml based on a Puppetfile, the modules in a
+  directory, or the combination of both.
+  
 ### 5.4.3 / 2018-03-29
 * Fix RPM release processing when generating 'Obsoletes' metadata